backend/requirements.txt
asgiref==3.6.0
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
flake8==5.0.4
Pillow==9.4.0
psycopg2-binary==2.8.6
pytz==2022.7.1
requests==2.26.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
